+++ docs/changelog.md
@@ -2,6 +2,13 @@
 
 ----
 
+## v2.9.1 (2024-01-11)
+
+- Inlined one in CoreParserTest.php because it was used only in one place
+- Change IfStatement ast node to a better implementation
+
+----
+
 ## v2.9.0 (2023-12-23)
 
 - Simplified Lexer code